+++ Chapter03/sumMatrix.cu
@@ -73,10 +73,10 @@ int main(int argc, char **argv) {
     gpuRef = (float *) malloc(nBytes);
 
     // initialize data at host side
-    size_t iStart = seconds();
+    double iStart = seconds();
     initialData(h_A, nxy);
     initialData(h_B, nxy);
-    size_t iElaps = seconds() - iStart;
+    double iElaps = seconds() - iStart;
 
     memset(hostRef, 0, nBytes);
     memset(gpuRef, 0, nBytes);
@@ -114,7 +114,7 @@ int main(int argc, char **argv) {
     sumMatrixOnGPU2D<<<grid, block>>>(d_MatA, d_MatB, d_MatC, nx, ny);
     CHECK(cudaDeviceSynchronize());
     iElaps = seconds() - iStart;
-    printf("sumMatrixOnGPU2D <<<(%d,%d), (%d,%d)>>> elapsed %zu ms\n", grid.x,
+    printf("sumMatrixOnGPU2D <<<(%d,%d), (%d,%d)>>> elapsed %f ms\n", grid.x,
            grid.y,
            block.x, block.y, iElaps);
     CHECK(cudaGetLastError());
